Google Cloud Functions/requirements.txt
functions-framework==3.*
gcloud
google-cloud-storage
